Reconnect with POE::Component::IRC::Plugin::Connector
Removed all the reconnection code and loaded the Connector plugin
instead. The old code would reconnect everytime we got disconnected,
making it impossible to shut down the bot cleanly.

Documented the proper way to shut it down.

Bumped the POE::Component::IRC dependency to 6.35 which fixes a
long-standing bug in the Connector plugin.