Disable bogus NAMES query
I have no idea what this was for. At this point it has already received
the NAMES replies, so why is it calling the "names" method, which
AUTOLOADs to sending a NAMES query to the IRC server? Also, it's doing
so with a hashref as an argument, resulting in "NAMES HASH(0x1234567)"